Created service/simulator package and moved base/simulators/ABCSimulator there. Old base/simulators was renamed to base/simulation_settings.
