+++ tvb_epilepsy/custom/simulator_custom.py
@@ -13,10 +13,11 @@ from copy import copy
 import numpy
 from tvb_epilepsy.base.constants.module_constants import LIB_PATH, HDF5_LIB, JAR_PATH, JAVA_MAIN_SIM
 from tvb_epilepsy.base.utils.log_error_utils import warning
-from tvb_epilepsy.base.utils.data_structures_utils import obj_to_dict, assert_arrays, construct_import_path
-from tvb_epilepsy.base.simulators import ABCSimulator, SimulationSettings
+from tvb_epilepsy.base.utils.data_structures_utils import obj_to_dict, assert_arrays
+from tvb_epilepsy.base.simulation_settings import SimulationSettings
 from tvb_epilepsy.base.computations.calculations_utils import calc_x0_val_to_model_x0
 from tvb_epilepsy.io.h5_reader import H5Reader
+from tvb_epilepsy.service.simulator.simulator import ABCSimulator
 
 
 class SimulationSettings(object):
