+++ tvb_epilepsy/service/pse_service.py
@@ -11,7 +11,7 @@ from tvb_epilepsy.base.utils.log_error_utils import initialize_logger, warning, 
 from tvb_epilepsy.base.utils.data_structures_utils import formal_repr
 from tvb_epilepsy.base.model.disease_hypothesis import DiseaseHypothesis
 from tvb_epilepsy.base.model.model_configuration import ModelConfiguration
-from tvb_epilepsy.base.simulators import ABCSimulator
+from tvb_epilepsy.service.simulator.simulator import ABCSimulator
 from tvb_epilepsy.service.epileptor_model_factory import model_build_dict
 from tvb_epilepsy.service.model_configuration_service import ModelConfigurationService
 from tvb_epilepsy.service.lsa_service import LSAService
