+++ tvb_epilepsy/tests/test_custom_h5_reader.py
@@ -4,7 +4,7 @@ from tvb_epilepsy.base.constants.configurations import DATA_TEST
 from tvb_epilepsy.base.model.disease_hypothesis import DiseaseHypothesis
 from tvb_epilepsy.base.model.model_configuration import ModelConfiguration
 from tvb_epilepsy.base.model.vep.sensors import Sensors
-from tvb_epilepsy.base.simulators import SimulationSettings
+from tvb_epilepsy.base.simulation_settings import SimulationSettings
 from tvb_epilepsy.io.h5_reader import H5Reader
 from tvb_epilepsy.io.h5_writer import H5Writer
 from tvb_epilepsy.service.lsa_service import LSAService
